# docs/API_CONTRACT.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Talas API",
		"description": "API Contract untuk Project Talas - Platform showcase project untuk developer & desainer dengan fitur sosial microblogging.",
		"version": "1.0.0",
		"contact": {
			"name": "Talas Team"
		}
	},
	"servers": [
		{
			"url": "http://localhost:80/api",
			"description": "Local Development (via Nginx Gateway)"
		},
		{
			"url": "http://localhost:3001/api",
			"description": "User Service Direct"
		},
		{
			"url": "http://localhost:8080/api",
			"description": "Content Service Direct"
		}
	],
...
	"paths": {
...
		"/showcases/{id}/comments": {
			"get": {
				"tags": ["Comments"],
				"summary": "Get comments for showcase",
				"description": "Mendapatkan daftar komentar pada showcase. **Endpoint ini PUBLIC**. Response berisi nested comments (reply tanpa batas level).",
				"operationId": "getShowcaseComments",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "UUID showcase",
						"schema": {
							"type": "string",
							"format": "uuid",
							"example": "550e8400-e29b-41d4-a716-446655440002"
						}
					},
					{
						"name": "cursor",
						"in": "query",
						"required": false,
						"description": "Cursor untuk pagination",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"description": "Jumlah komentar per halaman (default: 20)",
						"schema": {
							"type": "integer",
							"default": 20,
							"example": 20
						}
					}
				],
				"responses": {
					"200": {
						"description": "Daftar komentar berhasil diambil",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentsResponse"
								},
								"example": {
									"code": 200,
									"success": true,
									"data": {
										"comments": [
											{
												"id": "550e8400-e29b-41d4-a716-446655440020",
												"parent_id": null,
												"content": "Amazing project! Love the design.",
												"count_like": 5,
												"is_liked": false,
												"created_at": "2026-01-27T05:00:00Z",
												"author": {
													"id": "550e8400-e29b-41d4-a716-446655440010",
													"username": "janedoe",
													"name": "Jane Doe",
													"avatar_url": "https://res.cloudinary.com/talas/avatar2.jpg"
												},
												"replies": [
													{
														"id": "550e8400-e29b-41d4-a716-446655440021",
														"parent_id": "550e8400-e29b-41d4-a716-446655440020",
														"content": "Thanks! Appreciate the feedback.",
														"reply_to": "janedoe",
														"count_like": 2,
														"is_liked": false,
														"created_at": "2026-01-27T05:30:00Z",
														"author": {
															"id": "550e8400-e29b-41d4-a716-446655440000",
															"username": "johndoe",
															"name": "John Doe",
															"avatar_url": "https://res.cloudinary.com/talas/avatar.jpg"
														},
														"replies": []
													}
												]
											}
										],
										"pagination": {
											"next_cursor": "eyJpZCI...",
											"has_next": true
										}
									},
									"errors": null
								}
							}
						}
					},
					"404": {
						"description": "Showcase tidak ditemukan",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								},
								"example": {
									"code": 404,
									"success": false,
									"data": null,
									"errors": [
										{
											"field": "id",
											"message": "Showcase tidak ditemukan"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"example": {
									"code": 500,
									"success": false,
									"data": null,
									"errors": [
										{
											"field": "server",
											"message": "Terjadi kesalahan pada server"
										}
									]
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Comments"],
				"summary": "Create comment on showcase",
				"description": "Membuat komentar baru pada showcase. **Requires Cookie Session (JWT)**. Untuk reply, sertakan parent_id (id komentar yang ingin di-reply).",
				"operationId": "createShowcaseComment",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "UUID showcase",
						"schema": {
							"type": "string",
							"format": "uuid",
							"example": "550e8400-e29b-41d4-a716-446655440002"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentRequest"
							},
							"examples": {
								"top_level": {
									"summary": "Komentar top-level",
									"value": {
										"content": "Great work! Really inspiring project."
									}
								},
								"reply": {
									"summary": "Reply ke komentar lain",
									"value": {
										"content": "I agree, the design is amazing!",
										"parent_id": "550e8400-e29b-41d4-a716-446655440020"
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Komentar berhasil dibuat",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								},
								"example": {
									"code": 201,
									"success": true,
									"data": {
										"comment": {
											"id": "550e8400-e29b-41d4-a716-446655440030",
											"parent_id": null,
											"project_id": "550e8400-e29b-41d4-a716-446655440002",
											"content": "Great work! Really inspiring project.",
											"count_like": 0,
											"created_at": "2026-01-27T06:00:00Z",
											"author": {
												"id": "550e8400-e29b-41d4-a716-446655440010",
												"username": "janedoe",
												"name": "Jane Doe",
												"avatar_url": "https://res.cloudinary.com/talas/avatar2.jpg"
											}
										}
									},
									"errors": null
								}
							}
						}
					},
					"400": {
						"description": "Validasi gagal",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								},
								"examples": {
									"empty_content": {
										"summary": "Content wajib diisi",
										"value": {
											"code": 400,
											"success": false,
											"data": null,
											"errors": [
												{
													"field": "content",
													"message": "Content wajib diisi"
												}
											]
										}
									},
									"invalid_parent": {
										"summary": "Parent comment tidak ditemukan",
										"value": {
											"code": 400,
											"success": false,
											"data": null,
											"errors": [
												{
													"field": "parent_id",
													"message": "Komentar parent tidak ditemukan"
												}
											]
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Tidak terautentikasi",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								},
								"example": {
									"code": 401,
									"success": false,
									"data": null,
									"errors": [
										{
											"field": "auth",
											"message": "Token tidak valid atau sudah expired"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "Showcase tidak ditemukan",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								},
								"example": {
									"code": 404,
									"success": false,
									"data": null,
									"errors": [
										{
											"field": "id",
											"message": "Showcase tidak ditemukan"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								},
								"example": {
									"code": 500,
									"success": false,
									"data": null,
									"errors": [
										{
											"field": "server",
											"message": "Terjadi kesalahan pada server"
										}
									]
								}
							}
						}
					}
				}
			}
		},
		"/comments/{id}": {
			"patch": {
				"tags": ["Comments"],
				"summary": "Update comment",
				"description": "Edit konten komentar. **Requires Cookie Session (JWT)**. Hanya pemilik komentar yang dapat mengedit.",
				"operationId": "updateComment",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "UUID komentar yang akan diupdate",
						"schema": {
							"type": "string",
							"format": "uuid",
							"example": "550e8400-e29b-41d4-a716-446655440020"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCommentRequest"
							},
							"example": {
								"content": "Amazing project! Love the design and attention to detail."
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Komentar berhasil diupdate",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateCommentResponse"
								}
							}
						}
					},
					"400": {
						"description": "Validasi gagal",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden (Bukan pemilik komentar)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Komentar tidak ditemukan",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Comments"],
				"summary": "Delete comment",
				"description": "Hapus komentar (Soft delete). **Requires Cookie Session (JWT)**. Hanya pemilik komentar yang dapat menghapus. Jika punya child, konten diganti jadi tombstone.",
				"operationId": "deleteComment",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "UUID komentar",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Komentar berhasil dihapus",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteCommentResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/comments/{id}/reply": {
			"post": {
				"tags": ["Comments"],
				"summary": "Reply to comment",
				"description": "Membalas komentar. **Requires Cookie Session (JWT)**. ID pada URL adalah parent comment ID.",
				"operationId": "replyComment",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "UUID parent comment",
						"schema": {
							"type": "string",
							"format": "uuid",
							"example": "550e8400-e29b-41d4-a716-446655440020"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReplyCommentRequest"
							},
							"example": {
								"content": "Thanks for the feedback! I appreciate it."
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Berhasil membalas komentar",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentReplyResponse"
								},
								"example": {
									"code": 201,
									"success": true,
									"data": {
										"comment": {
											"id": "550e8400-e29b-41d4-a716-446655440055",
											"parent_id": "550e8400-e29b-41d4-a716-446655440020",
											"project_id": "550e8400-e29b-41d4-a716-446655440002",
											"content": "Thanks for the feedback! I appreciate it.",
											"reply_to": "janedoe",
											"count_like": 0,
											"is_liked": false,
											"created_at": "2026-01-27T08:00:00Z",
											"updated_at": "2026-01-27T08:00:00Z",
											"author": {
												"id": "550e8400-e29b-41d4-a716-446655440010",
												"username": "janedoe",
												"name": "Jane Doe",
												"avatar_url": "https://res.cloudinary.com/talas/avatar2.jpg"
											}
										}
									},
									"errors": null
								}
							}
						}
					},
					"400": {
						"description": "Validasi gagal",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Tidak terautentikasi",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Parent comment tidak ditemukan",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/comments/{id}/like": {
			"post": {
				"tags": ["Comments"],
				"summary": "Toggle like on comment",
				"description": "Like atau Unlike komentar. **Requires Cookie Session (JWT)**. Toggle behavior: jika sudah like akan unlike, jika belum like akan like.",
				"operationId": "toggleCommentLike",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						},
						"description": "Comment UUID"
					}
				],
				"responses": {
					"200": {
						"description": "Berhasil toggle like",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"code": { "type": "integer", "example": 200 },
										"success": { "type": "boolean", "example": true },
										"message": {
											"type": "string",
											"example": "Success toggle like"
										},
										"data": {
											"type": "object",
											"properties": {
												"liked": {
													"type": "boolean",
													"description": "true jika baru saja di-like, false jika baru saja di-unlike"
												},
												"likes_count": {
													"type": "integer",
													"description": "Jumlah total likes pada comment"
												}
											}
										}
									}
								},
								"example": {
									"code": 200,
									"success": true,
									"message": "Success toggle like",
									"data": {
										"liked": true,
										"likes_count": 5
									}
								}
							}
						}
					},
					"404": {
						"description": "Comment tidak ditemukan",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"ApiResponse": {
				"type": "object",
				"description": "Standard API response wrapper",
				"properties": {
					"code": {
						"type": "integer",
						"description": "HTTP status code",
						"example": 200
					},
					"success": {
						"type": "boolean",
						"description": "Indikator keberhasilan request",
						"example": true
					},
					"data": {
						"type": "object",
						"nullable": true,
						"description": "Data response (null jika error)"
					},
					"errors": {
						"type": "array",
						"nullable": true,
						"description": "Array of error objects (null jika success)",
						"items": {
							"$ref": "#/components/schemas/FieldError"
						}
					}
				},
				"required": ["code", "success"]
			},
...
			"CommentsResponse": {
				"type": "object",
				"description": "Response untuk GET /showcases/:id/comments",
				"properties": {
					"code": {
						"type": "integer",
						"example": 200
					},
					"success": {
						"type": "boolean",
						"example": true
					},
					"data": {
						"type": "object",
						"properties": {
							"comments": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Comment"
								}
							},
							"pagination": {
								"$ref": "#/components/schemas/CursorPaginationMeta"
							}
						}
					},
					"errors": {
						"type": "array",
						"nullable": true,
						"items": {
							"$ref": "#/components/schemas/FieldError"
						}
					}
				}
			},
...
			"ErrorResponse": {
				"type": "object",
				"description": "Standard error response",
				"properties": {
					"code": {
						"type": "integer",
						"description": "HTTP status code"
					},
					"success": {
						"type": "boolean",
						"example": false
					},
					"data": {
						"type": "object",
						"nullable": true,
						"example": null
					},
					"errors": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/FieldError"
						}
					}
				}
			},
...
			"UpdateCommentRequest": {
				"type": "object",
				"description": "Request body untuk update comment",
				"required": ["content"],
				"properties": {
					"content": {
						"type": "string",
						"description": "Isi komentar yang sudah diedit",
						"minLength": 1,
						"maxLength": 1000,
						"example": "Amazing project! Love the design and attention to detail."
					}
				}
			},
...
		},
		"securitySchemes": {
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "access_token",
				"description": "JWT access token disimpan dalam HttpOnly cookie. Dikirim otomatis oleh browser."
			},
			"refreshTokenAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "refresh_token",
				"description": "JWT refresh token disimpan dalam HttpOnly cookie. Digunakan untuk mendapatkan access token baru."
			}
		},
...
	}
}